mcp-servers/diagnose/requirements.txt
google-adk==1.8.0
nest_asyncio
toolbox-core